enh: enh for cloud-tdgpt

+++ enterprise/packaging/docker/dockerbuild.sh
@@ -17,12 +17,14 @@ cpuTypeAlias=""
 version=""
 passWord=""
 pkgFile=""
+tdgptPkgFile=""
 verType="stable"
 dockerLatest="n"
 cloudBuild="n"
 dockerProject="tdengine"
+nasIp="0.0.0.0"
 
-while getopts "hc:n:p:f:V:a:b:d:D:" arg
+while getopts "hc:n:p:f:V:a:b:d:D:g:i:" arg
 do
   case $arg in
     c)
@@ -48,6 +50,14 @@ do
     V)
       #echo "verType=$OPTARG"
       verType=$(echo $OPTARG)
+      ;;
+    g)
+      #echo "dockerLatest=$OPTARG"
+      tdgptPkgFile=$(echo $OPTARG)
+      ;;
+    i)
+      #echo "dockerLatest=$OPTARG"
+      nasIp=$(echo $OPTARG)
       ;;
     d)
       #echo "cloudBuild=$OPTARG"
@@ -88,9 +98,15 @@ done
 if [ "$verType" == "beta" ]; then
   dockername=${cpuType}-${verType}
   dirName=${pkgFile%-beta*}
+  if [ "$tdGpt" == "true" ];then
+    tdgptDirName=${tdgptPkgFile%-beta*}
+  fi
 elif [ "$verType" == "stable" ]; then
   dockername=${cpuType}
   dirName=${pkgFile%-Linux*}
+  if [ "$tdGpt" == "true" ];then
+    tdgptDirName=${tdgptPkgFile%-Linux*}
+  fi
 else
   echo "unknown verType, nor stabel or beta"
   exit 1
@@ -109,7 +125,11 @@ enterpriseDir=${scriptDir}/../../../enterprise
 DockerfilePath=${enterpriseDir}/packaging/docker/
 if [ "$cloudBuild" == "y" ]; then
   communityArchiveDir=/nas/TDengine/v$version/cloud
-  Dockerfile=${enterpriseDir}/packaging/docker/DockerfileCloud
+  if [ "$tdGpt" == "true" ];then
+    Dockerfile=${enterpriseDir}/packaging/docker/DockerfileCloudTDgpt
+  else
+    Dockerfile=${enterpriseDir}/packaging/docker/DockerfileCloud
+  fi
 else
   communityArchiveDir=/nas/TDengine/v$version/community
   Dockerfile=${enterpriseDir}/packaging/docker/Dockerfile
@@ -130,7 +150,11 @@ else
     exit 1
 fi
 
-docker build --rm -f "${Dockerfile}"  --network=host -t ${dockerProject}/tdengine-enterprise-${dockername}:${version} "." --build-arg pkgFile=${pkgFile} --build-arg dirName=${dirName} --build-arg cpuType=${cpuTypeAlias}
+if [ "$tdGpt" == "true" ];then
+  docker build --rm -f "${Dockerfile}"  --network=host -t ${dockerProject}/tdengine-enterprise-${dockername}:${version} "." --build-arg pkgFile=${pkgFile}  --build-arg dirName=${dirName} --build-arg tdgptPkgFile=${tdgptPkgFile}  --build-arg tdgptDirName=${tdgptDirName}  --build-arg cpuType=${cpuTypeAlias} --build-arg nasIp=${nasIp}
+else
+  docker build --rm -f "${Dockerfile}"  --network=host -t ${dockerProject}/tdengine-enterprise-${dockername}:${version} "." --build-arg pkgFile=${pkgFile} --build-arg dirName=${dirName} --build-arg cpuType=${cpuTypeAlias}
+fi
 
 docker logout
 docker login https://image.cloud.taosdata.com -u internaltest -p ${passWord}  #replace the docker registry username and password
